requirements.txt
tensorflow-gpu==1.10.0
scikit-learn==0.19.1
numpy==1.14.4
scipy==1.1.0
pandas==0.23.0
opencv-python==3.4.2.17
tqdm==4.23.4
